requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.3
numpy>=1.19.5
pillow>=8.3.1
pandas>=1.3.0
requests>=2.26.0
kaggle>=1.5.12 